Made a C script which does the byte pair encoding for much faster speeds, but haven't finished implementing it with the python scripts.
Made a database using the module 'shelve' which can store python objects and be queried without loading into ram.

Made a script which goes through the comment DB and finds the parent comments+submissions, runs it through the BPE token converter thing and makes PyTorch tensors with the correct keys.